projeto-final/requirements.txt
click==8.1.3
Flask==2.1.2
ImageHash==4.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
Pillow==9.1.1
PyWavelets==1.3.0
scipy==1.8.1
six==1.16.0
Werkzeug==2.1.2
